update: ddd folder structure for backend and frontend
