+++ package.json
@@ -12,7 +12,12 @@
     "test:e2e:ui": "playwright test --ui",
     "generate:logos": "bash scripts/generate-logo-assets.sh",
     "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
     "format": "prettier --write \"src/**/*.{ts,tsx,js,jsx,css,json}\"",
+    "typecheck": "tsc --noEmit",
+    "rust:fmt": "cargo fmt --all --manifest-path src-tauri/Cargo.toml -- --check",
+    "rust:lint": "cargo clippy --all-targets --manifest-path src-tauri/Cargo.toml -- -D warnings",
+    "rust:check": "bun run rust:fmt && bun run rust:lint",
     "prepare": "husky"
   },
   "lint-staged": {
