+++ src-tauri/src/infrastructure/mod.rs
@@ -1,0 +1,2 @@
+pub mod bot_lifecycle;
+pub mod http_server;
