+++ src-tauri/src/modules/bot/mod.rs
@@ -1,0 +1,3 @@
+pub mod commands;
+pub mod repository;
+pub mod service;
