+++ src-tauri/src/modules/bot/repository.rs
@@ -1,0 +1,24 @@
+use crate::shared::state::ConnectionData;
+use std::path::Path;
+
+pub fn persist(path: &Path, data: &ConnectionData) -> Result<(), String> {
+    let json = serde_json::to_string_pretty(data).map_err(|e| e.to_string())?;
+    if let Some(parent) = path.parent() {
+        std::fs::create_dir_all(parent).map_err(|e| e.to_string())?;
+    }
+    std::fs::write(path, json).map_err(|e| e.to_string())?;
+    Ok(())
+}
+
+pub fn load(path: &Path) -> Option<ConnectionData> {
+    let json = std::fs::read_to_string(path).ok()?;
+    serde_json::from_str(&json).ok()
+}
+
+pub fn clear(path: &Path) -> Result<(), String> {
+    match std::fs::remove_file(path) {
+        Ok(()) => Ok(()),
+        Err(e) if e.kind() == std::io::ErrorKind::NotFound => Ok(()),
+        Err(e) => Err(e.to_string()),
+    }
+}
