+++ src-tauri/src/modules/mod.rs
@@ -1,0 +1,2 @@
+pub mod bot;
+pub mod ollama;
