+++ src-tauri/src/modules/ollama/constants.rs
@@ -1,0 +1,3 @@
+pub const OLLAMA_PS_URL: &str = "http://localhost:11434/api/ps";
+pub const OLLAMA_TAGS_URL: &str = "http://localhost:11434/api/tags";
+pub const OLLAMA_CHAT_URL: &str = "http://localhost:11434/api/chat";
