+++ src-tauri/src/modules/ollama/mod.rs
@@ -1,0 +1,2 @@
+pub mod constants;
+pub mod service;
