+++ src/modules/bot/index.ts
@@ -1,0 +1,5 @@
+export { getPengineHealth, postConnect, deleteConnect, PENGINE } from "./api";
+export type { PengineHealth } from "./types";
+export { useAppSessionStore } from "./store/appSessionStore";
+export { SetupWizard, SETUP_STEPS } from "./components/SetupWizard";
+export { TerminalPreview } from "./components/TerminalPreview";
