+++ src/modules/bot/types.ts
@@ -1,0 +1,6 @@
+export type PengineHealth = {
+  status: string;
+  bot_connected: boolean;
+  bot_username?: string;
+  bot_id?: string | null;
+};
