+++ src/modules/ollama/index.ts
@@ -1,0 +1,2 @@
+export { fetchOllamaModel } from "./api";
+export type { OllamaProbe } from "./types";
