+++ src/modules/ollama/types.ts
@@ -1,0 +1,1 @@
+export type OllamaProbe = { reachable: boolean; model: string | null };
