+++ src/shared/ui/PhoneMockup.tsx
