+++ src/shared/ui/SpecMockup.tsx
