+++ src/shared/ui/StyledQrCode.tsx
