+++ src/shared/ui/WizardLayout.tsx
